requirements.txt
certifi==2021.5.30
chardet==4.0.0
Django==2.2.10
django-nested-admin==3.3.3
django-rest-framework==0.1.0
djangorestframework==3.12.4
idna==2.10
python-monkey-business==1.0.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
